--- lib/database/entity.dart
@@ -1,4 +1,0 @@
-abstract class Entity {
-  Map<String, dynamic> toMap();
-  Entity fromMap(Map<String, dynamic> map);
-}
