+++ lib/models/abstract_entity/entity.dart
@@ -1,0 +1,4 @@
+abstract class AbstractEntity {
+  Map<String, dynamic> toMap();
+  AbstractEntity fromMap(Map<String, dynamic> map);
+}
