+++ test/unit/business_card_bloc_test.dart
@@ -1,6 +1,6 @@
 // Import the test package and Counter class
 import 'package:flutter_test/flutter_test.dart';
-import 'package:quicksnap/business_card/business_card_bloc.dart';
+import 'package:quicksnap/blocs/business_card_bloc.dart';
 
 void main() {
   group('Email Validation', () {
